+++ lib/deploy_config_test.go
@@ -10,10 +10,16 @@ func TestValidateRepair(t *testing.T) {
 	dc := DeployConfig{
 		Volumes:   Volumes{nil, &Volume{}},
 		Resources: make(Resources),
+		Startup: Startup{
+
+			SkipCheck: true,
+		},
 	}
 	dc.Resources["cpus"] = "0.25"
 	dc.Resources["memory"] = "356"
 	dc.Resources["ports"] = "2"
+
+	t.Log(dc.Startup)
 
 	assert.Len(t, dc.Volumes, 2)
 	flaws := dc.Validate()