+++ lib/state_test.go
@@ -48,6 +48,9 @@ func TestState_Validate(t *testing.T) {
 								"ports":  "1",
 							},
 							NumInstances: 3,
+							Startup: Startup{
+								SkipCheck: true,
+							},
 						},
 						Version: semv.MustParse("1"),
 					},